requirements.txt
pypdf2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.3
requests==2.31.0
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
schedule==1.2.1
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.18.1
scikit-learn==1.4.0
nltk==3.8.1
